refactor state implementations folder
