+++ src/lib/state/persistent/index.ts
@@ -1,4 +1,4 @@
 import { AbstractPersistentState } from "./abstract-persistent-state";
-import { SessionStorageState } from "./session-storage-state";
+import { SessionStorageState } from "./implementations/session-storage-state";
 
 export { AbstractPersistentState, SessionStorageState };
